Refactor: 마이 대쉬보드 폴더 구조 리팩토링
